# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Bapz Store Backend",
      "version": "1.0.0",
      "description": "Bapz Store Backend Swagger"
    },
    "basePath": "/",
    "schemes": [
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/api/bapz/id": {
        "post": {
          "tags": [
            "Bapz"
          ],
          "summary": "Retrieve a product by ID",
          "description": "retrieve product full details by ID",
          "parameters": [
            {
              "name": "id",
              "in": "body",
              "description": "The ID of the image",
              "required": true,
              "type": "object",
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "default":736
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Image uploaded successfully"
            },
            "400": {
              "description": "Invalid request data"
            }
          }
        }
      },
      "/images": {
        "get": {
          "tags": [
            "Bapz"
          ],
          "summary": "Retrieve all images",
          "description": "Get mapping of all images names and corresponding src files.",
          "responses": {
            "200": {
              "description": "Data ready."
            },
            "400": {
              "description": "Invalid request data"
            }
          }
        }
      },
      "/api/bapz/apparel": {
        "post": {
          "tags": [
            "Bapz"
          ],
          "summary": "Retrieve all images by category",
          "description": "Get mapping of all images data by category.If cat is not specified it will return 2 products of each categry",
          "parameters": [
            {
              "name": "cat",
              "in": "body",
              "description": "The category of products",
              "required": true,
              "type": "object",
              "schema": {
                "type": "object",
                "properties": {
                  "cat": {
                    "type": "string",
                    "default":"hats"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Data ready."
            },
            "400": {
              "description": "Invalid request data"
            }
          }
        }
      },
      "/ids": {
        "post": {
          "tags": [
            "Bapz"
          ],
          "summary": "Retrieve IDs",
          "description": "Get all product's ids",
          "parameters": [
            {
              "name": "limit",
              "in": "body",
              "description": "Maximum results returned",
              "required": false,
              "type": "object",
              "schema": {
                "type": "object",
                "properties": {
                  "limit": {
                    "type": "integer",
                    "default": "10"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "IDs ready."
            },
            "400": {
              "description": "Invalid request data"
            }
          }
        }
      },
      "/": {
        "get": {
          "tags": [
            "Bapz"
          ],
          "summary": "Test API is up",
          "responses": {
            "200": {
              "description": "API is up"
            },
            "400": {
              "description": "Invalid request data"
            }
          }
        }
      }
    }
  }
  